% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latitudelongitude.R
\name{lat_long}
\alias{lat_long}
\title{Retrieve Latitude and Longitude for a Given Location}
\usage{
lat_long(loca_tion)
}
\arguments{
\item{loca_tion}{A character string representing the location for which
latitude and longitude are to be retrieved.}
}
\value{
A list containing the latitude and longitude coordinates of the
  given location.
}
\description{
This function takes a location string, converts spaces to '+', and retrieves
latitude and longitude coordinates for that location using the Nominatim
API from OpenStreetMap.
}
\examples{
\dontrun{
  result <- lat_long("New York City")
  print(result)
}
}
