\name{lab5Package-package}
\alias{lab5Package-package}
\alias{lab5Package}
\docType{package}
\title{
\packageTitle{lab5Package}
}
\description{
\packageDescription{lab5Package}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{lab5Package}
\packageIndices{lab5Package}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{lab5Package}

Maintainer: \packageMaintainer{lab5Package}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS ~~
~~ in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
